address cr issues
